requirements.txt
python-dotenv
# For SQL Server access and Azure AD authentication
pyodbc
azure-identity
uv
fastmcp